Permissions System
Set up the basic concept behind how permissions will be handled for the bot. General idea is as follows:

Any command that requires a permission will have the element hasPermission: true as a part of their module exports.

As well it will need to have a PermissionsSettings object will be organize in the following way:

PermissionSettings: {
    "ROLE": {
        "Role ID": true of false (to allow or disallow)
     },
    "USER": {
        "UserID": true or false
    }
}

Either object can be empty. Commands that have permissions will be disallowed by default. Will most likely create a setting to change that at some point